clean up main.cs
